Split response code out into separate file and correct request tests
